fix typos fix vehicle pose not properly set force setting Item tag on EntityItem

+++ 0-KFCommonUtilityLib/KFAttached/Animation/MonoBehaviours/AnimatorWrapper/AnimatorWrapper.cs
@@ -119,7 +119,7 @@ public class AnimatorWrapper : IAnimatorWrapper
 
     public void SetFloat(int id, float value) => animator.SetFloat(id, value);
 
-    public void SetInteger(string name, int value) => animator.SetFloat(name, value);
+    public void SetInteger(string name, int value) => animator.SetInteger(name, value);
 
     public void SetInteger(int id, int value) => animator.SetInteger(id, value);
